Refactored repository/model location URI
