+++ src/main/java/com/raf/si/Banka2Backend/repositories/mariadb/PermissionRepository.java
@@ -1,7 +1,7 @@
-package com.raf.si.Banka2Backend.repositories.users;
+package com.raf.si.Banka2Backend.repositories.mariadb;
 
-import com.raf.si.Banka2Backend.models.users.Permission;
-import com.raf.si.Banka2Backend.models.users.PermissionName;
+import com.raf.si.Banka2Backend.models.mariadb.Permission;
+import com.raf.si.Banka2Backend.models.mariadb.PermissionName;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
