+++ src/main/java/com/raf/si/Banka2Backend/repositories/mongodb/ForexRepository.java
@@ -1,6 +1,6 @@
-package com.raf.si.Banka2Backend.repositories.exchange;
+package com.raf.si.Banka2Backend.repositories.mongodb;
 
-import com.raf.si.Banka2Backend.models.exchange.Forex;
+import com.raf.si.Banka2Backend.models.mongodb.Forex;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface ForexRepository extends MongoRepository<Forex, String> {}
