+++ src/main/java/com/raf/si/Banka2Backend/requests/RegisterRequest.java
@@ -1,6 +1,6 @@
 package com.raf.si.Banka2Backend.requests;
 
-import com.raf.si.Banka2Backend.models.users.PermissionName;
+import com.raf.si.Banka2Backend.models.mariadb.PermissionName;
 import java.util.List;
 import lombok.Data;
