+++ src/main/java/com/raf/si/Banka2Backend/services/AuthorisationService.java
@@ -1,12 +1,12 @@
 package com.raf.si.Banka2Backend.services;
 
-import com.raf.si.Banka2Backend.models.users.PasswordResetToken;
-import com.raf.si.Banka2Backend.models.users.Permission;
-import com.raf.si.Banka2Backend.models.users.PermissionName;
-import com.raf.si.Banka2Backend.models.users.User;
-import com.raf.si.Banka2Backend.repositories.users.PasswordResetTokenRepository;
-import com.raf.si.Banka2Backend.repositories.users.PermissionRepository;
-import com.raf.si.Banka2Backend.repositories.users.UserRepository;
+import com.raf.si.Banka2Backend.models.mariadb.PasswordResetToken;
+import com.raf.si.Banka2Backend.models.mariadb.Permission;
+import com.raf.si.Banka2Backend.models.mariadb.PermissionName;
+import com.raf.si.Banka2Backend.models.mariadb.User;
+import com.raf.si.Banka2Backend.repositories.mariadb.PasswordResetTokenRepository;
+import com.raf.si.Banka2Backend.repositories.mariadb.PermissionRepository;
+import com.raf.si.Banka2Backend.repositories.mariadb.UserRepository;
 import java.util.Calendar;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
