+++ src/main/java/com/raf/si/Banka2Backend/services/interfaces/PermissionServiceInterface.java
@@ -1,7 +1,7 @@
 package com.raf.si.Banka2Backend.services.interfaces;
 
-import com.raf.si.Banka2Backend.models.users.Permission;
-import com.raf.si.Banka2Backend.models.users.PermissionName;
+import com.raf.si.Banka2Backend.models.mariadb.Permission;
+import com.raf.si.Banka2Backend.models.mariadb.PermissionName;
 import java.util.List;
 
 public interface PermissionServiceInterface {
