+++ src/test/java/com/raf/si/Banka2Backend/services/AuthorisationServiceTest.java
@@ -4,12 +4,12 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.raf.si.Banka2Backend.models.users.PasswordResetToken;
-import com.raf.si.Banka2Backend.models.users.Permission;
-import com.raf.si.Banka2Backend.models.users.PermissionName;
-import com.raf.si.Banka2Backend.models.users.User;
-import com.raf.si.Banka2Backend.repositories.users.PasswordResetTokenRepository;
-import com.raf.si.Banka2Backend.repositories.users.UserRepository;
+import com.raf.si.Banka2Backend.models.mariadb.PasswordResetToken;
+import com.raf.si.Banka2Backend.models.mariadb.Permission;
+import com.raf.si.Banka2Backend.models.mariadb.PermissionName;
+import com.raf.si.Banka2Backend.models.mariadb.User;
+import com.raf.si.Banka2Backend.repositories.mariadb.PasswordResetTokenRepository;
+import com.raf.si.Banka2Backend.repositories.mariadb.UserRepository;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
